+++ abm.py
@@ -11,19 +11,13 @@ import yaml
 import sys
 import os
 import logging
-import common
-from common import parse_profile
+from lib import common
+from lib.common import parse_profile
 from pprint import pprint
 
 # These imports are required because they need to be added to the symbol table
 # so the parse_menu method can find them in globals()
-import job
-import dataset
-import workflow
-import history
-import library
-import folder
-import config
+from lib import job, dataset, workflow, history, library, folder
 
 log = logging.getLogger('abm')
 log.setLevel(logging.ERROR)
@@ -126,7 +120,7 @@ def alias(shortcut, fullname):
 
 def parse_menu():
     log.debug('parse_menu')
-    menu_config = f'{os.path.dirname(__file__)}/menu.yml'
+    menu_config = f'{os.path.dirname(__file__)}/lib/menu.yml'
     if not os.path.exists(menu_config):
         print(f"ERROR: Unable to load the menu configuration from {menu_config}")
         sys.exit(1)
@@ -161,7 +155,7 @@ def version():
     print(f"    Copyright 2021 The Galaxy Project. All Rights Reserved.\n")
 
 
-def main():
+def entrypoint():
     menu_data = parse_menu()
 
     if len(sys.argv) < 2 or sys.argv[1] in help_args:
@@ -227,4 +221,4 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    entrypoint()
