+++ lib/common.py
